Create postman-setup.md
Postman is an invaluable resource for new developers on this api. Having a pre-configured json file for postman will allow developers to understand exactly what a request looks like, and how they can make that request. Another option would be to host that json config file somewhere in this repo, so that we can make updates as we develop new endpoints.